Folder structure creation
Added folder in /main for each component.

In every folder there is CMakeLists.txt file (it is not required, but it is nice to have it in order to compile each subfolder on it's ow).

In CMakeLists.txt file is biggest change, added paths to subfolders + added all files from subcomponents to be compiled.